# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0",
        "title": "Survey Service",
        "description": "Microservice for designing and administering surveys.",
        "contact": {
            "name": "Amida Technology Solutions",
            "url": "http://www.amida.com",
            "email": "hello@amida-tech.com"
        },
        "license": {
            "name": "Apache 2.0"
        }
    },
    "basePath": "/api/v1.0",
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/answers/csv": {
            "x-swagger-router-controller": "answer.controller",
            "get": {
                "summary": "Exports answers as csv file",
                "operationId": "exportAnswers",
                "produces": ["text/csv", "application/json"],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "default": {
                        "$ref": "#/responses/genericError"
                    }
                }
            },
            "post": {
                "summary": "Imports answer csv file",
                "operationId": "importAnswers",
                "consumes": ["multipart/form-data"],
                "parameters": [{
                    "name": "answercsv",
                    "in": "formData",
                    "description": "Answer CSV file",
                    "required": true,
                    "type": "file"
                }, {
                    "name": "questionidmap",
                    "in": "formData",
                    "description": "Question Id Map",
                    "required": true,
                    "type": "string"
                }, {
                    "name": "surveyidmap",
                    "in": "formData",
                    "description": "Question Id Map",
                    "required": true,
                    "type": "string"
                }],
                "responses": {
                    "204": {
                        "description": "Successful response"
                    },
                    "default": {
                        "$ref": "#/responses/genericError"
                    }
                }
            }
        },
        "/answers/multi-user-csv": {
            "x-swagger-router-controller": "answer.controller",
            "get": {
                "summary": "Exports answers as csv file",
                "operationId": "exportMultiUserAnswers",
                "produces": ["text/csv", "application/json"],
                "parameters": [{
                    "name": "user-ids",
                    "in": "query",
                    "required": true,
                    "type": "array",
                    "items": {
                        "type": "integer",
                        "minimum": 1
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "default": {
                        "$ref": "#/responses/genericError"
                    }
                }
            },
            "post": {
                "summary": "Imports answer csv file",
                "operationId": "importAnswers",
                "consumes": ["multipart/form-data"],
                "parameters": [{
                    "name": "answercsv",
                    "in": "formData",
                    "description": "Answer CSV file",
                    "required": true,
                    "type": "file"
                }, {
                    "name": "questionidmap",
                    "in": "formData",
                    "description": "Question Id Map",
                    "required": true,
                    "type": "string"
                }, {
                    "name": "surveyidmap",
                    "in": "formData",
                    "description": "Question Id Map",
                    "required": true,
                    "type": "string"
                }, {
                    "name": "useridmap",
                    "in": "formData",
                    "description": "User Id Map",
                    "required": false,
                    "type": "string"
                }],
                "responses": {
                    "204": {
                        "description": "Successful response"
                    },
                    "default": {
                        "$ref": "#/responses/genericError"
                    }
                }
            }
        },
...
       "/assessment-answers": {
            "x-swagger-router-controller": "assessment-answer.controller",
            "get": {
                "summary": "Lists assessment answers",
                "description": "Do not return attached surveys.",
                "operationId": "getAssessmentAnswersList",
                "parameters": [{
                    "name": "assessment-answers-status",
                    "in": "query",
                    "description": "status of assessment completion by user",
                    "required": false,
                    "type":"string",
                    "enum":["new", "in-progress", "completed"]
                },{
                    "name": "group",
                    "in": "query",
                    "description": "Group for assessments",
                    "required": false,
                    "type": "string",
                    "minLength": 1
                }],
                "responses": {
                    "200": {
                        "description": "Successful response.",
                        "schema": {
                            "$ref": "#/definitions/assessmentAnswersDashboard"
                        }
                    },
                    "default": {
                        "$ref": "#/responses/genericError"
                    }
                }
            }
        },
        "/assessment-answers/{id}": {
            "x-swagger-router-controller": "assessment-answer.controller",
            "parameters": [{
                "name": "id",
                "in": "path",
                "required": true,
                "type": "integer"
            }],
            "post": {
                "description": "This persists assessment answers.  Old answers are soft deleted.  For each question for which there is no answer property, any existing answer will be soft deleted.",
                "summary": "Creates assessment answers",
                "operationId": "createAssessmentAnswers",
                "parameters": [{
                    "name": "newAssessmentAnswers",
                    "in": "body",
                    "description": "Answers to the survey",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/newAssessmentAnswers"
                    }
                }],
                "responses": {
                    "204": {
                        "description": "Successful response."
                    },
                    "default": {
                        "$ref": "#/responses/genericError"
                    }
                }
            },
            "get": {
                "summary": "Gets assessment answers",
                "operationId": "getAssessmentAnswers",
                "responses": {
                    "200": {
                        "description": "Successful response.",
                        "schema": {
                            "$ref": "#/definitions/assessmentAnswers"
                        }
                    },
                    "default": {
                        "$ref": "#/responses/genericError"
                    }
                }
            }
        },
        "/assessment-answers/{id}/as-copy": {
            "x-swagger-router-controller": "assessment-answer.controller",
            "parameters": [{
                "name": "id",
                "in": "path",
                "required": true,
                "type": "integer"
            }],
            "post": {
                "description": "This persists assessment answers.  Old answers are soft deleted.  For each question for which there is no answer property, any existing answer will be soft deleted.",
                "summary": "Copies assessment answers",
                "operationId": "copyAssessmentAnswers",
                "parameters": [{
                    "name": "prevAssessment",
                    "in": "body",
                    "description": "Assessment id to copy from",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/prevAssessment"
                    }
                }],
                "responses": {
                    "204": {
                        "description": "Successful response."
                    },
                    "default": {
                        "$ref": "#/responses/genericError"
                    }
                }
            }
        },
        "/assessment-answers/{id}/answers": {
            "x-swagger-router-controller": "assessment-answer.controller",
            "parameters": [{
                "name": "id",
                "in": "path",
                "required": true,
                "type": "integer"
            }],
            "get": {
                "summary": "Gets assessment answers only",
                "operationId": "getAssessmentAnswersOnly",
                "responses": {
                    "200": {
                        "description": "Successful response.",
                        "schema": {
                            "$ref": "#/definitions/assessmentLinkedAnswers"
                        }
                    },
                    "default": {
                        "$ref": "#/responses/genericError"
                    }
                }
            }
        },
        "/assessment-answers/csv" : {
            "x-swagger-router-controller": "assessment-answer.controller",
            "parameters":[{
                "description": "survey Id that of the assessments' answers that will be returned",
                "name": "survey-id",
                "in": "query",
                "required": true,
                "type":"integer"
            }, {
                "description": "questionId filter on the assessments' anwers that will be returned",
                "name": "question-id",
                "in": "query",
                "required": false,
                "type":"integer"
            },{
                "description": "sectionId filter on the assessments' answers that will be returned",
                "name": "section-id",
                "in": "query",
                "required": false,
                "type":"integer"
            }],
            "get": {
                "summary":"returns the answers of the question specified from latest assessments that use the specified survey, in json format",
                "operationId":"assessmentAnswersCSV",
                "produces": ["text/csv", "application/json"],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "default": {
                        "$ref": "#/responses/genericError"
                    }
                }

            },
            "post": {
                "summary": "Imports assessment answer csv file",
                "operationId": "importAssessmentAnswers",
                "consumes": ["multipart/form-data"],
                "parameters": [{
                    "name": "assessmentAnswercsv",
                    "in": "formData",
                    "description": "Answer CSV file",
                    "required": true,
                    "type": "file"
                }, {
                    "name": "questionidmap",
                    "in": "formData",
                    "description": "Question Id Map",
                    "required": true,
                    "type": "string"
                }, {
                    "name": "surveyidmap",
                    "in": "formData",
                    "description": "Question Id Map",
                    "required": true,
                    "type": "string"
                }, {
                    "name": "useridmap",
                    "in": "formData",
                    "description": "User Id Map",
                    "required": true,
                    "type": "string"
                }],
                "responses": {
                    "204": {
                        "description": "Successful response"
                    },
                    "default": {
                        "$ref": "#/responses/genericError"
                    }
                }
            }
        },
        "/assessment-answers/export" : {
            "x-swagger-router-controller": "assessment-answer.controller",
            "parameters":[{
                "description": "survey Id that of the assessments' answers that will be returned",
                "name": "survey-id",
                "in": "query",
                "required": true,
                "type":"integer"
            }, {
                "description": "questionId filter on the assessments' answers that will be returned",
                "name": "question-id",
                "in": "query",
                "required": false,
                "type":"integer"
            },{
                "description": "sectionId filter on the assessments' answers that will be returned",
                "name": "section-id",
                "in": "query",
                "required": false,
                "type":"integer"
            }, {
                "description": "boolean param to request comments along with question answers",
                "name": "include-comments",
                "in": "query",
                "required": false,
                "type":"boolean"
            }],
            "get": {
                "summary":"returns the answers of the question specified from latest assessments that use the specified survey, in csv format",
                "operationId":"exportAssessmentAnswers",
                "produces": ["application/json"],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "$ref": "#/definitions/assessmentAnswersExportList"
                        }
                    },
                    "default": {
                        "$ref": "#/responses/genericError"
                    }
                }
            }
        },
        "/assessment-answers/{id}/status": {
            "x-swagger-router-controller": "assessment-answer.controller",
            "parameters": [{
                "name": "id",
                "in": "path",
                "required": true,
                "type": "integer"
            }],
            "get": {
                "summary": "Gets assessment answers Status",
                "operationId": "getAssessmentAnswersStatus",
                "responses": {
                    "200": {
                        "description": "Successful response.",
                        "schema": {
                            "type": "object",
                            "required": ["status"],
                            "properties": {
                                "status": {
                                    "$ref": "#/definitions/answerStatus"
                                }
                            },
                            "additionalProperties": false
                        }
                    },
                    "default": {
                        "$ref": "#/responses/genericError"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "self": {
            "type": "apiKey",
            "in": "header",
            "name": "Authorization"
        }
    },
    "responses": {
        "genericError": {
            "description": "Error response",
            "schema": {
                "$ref": "#/definitions/error"
            }
        },
...
    },
    "definitions": {
...
        "error": {
            "type": "object",
            "required": [
                "message"
            ],
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "choicesItem": {
            "type": "object",
            "required": ["id"],
            "properties": {
                "id": {
                    "type": "integer",
                    "minimum": 1
                },
                "boolValue": {
                    "type": "boolean"
                },
                "integerValue": {
                    "type": "integer"
                },
                "floatValue": {
                    "type": "number"
                },
                "textValue": {
                    "type": "string"
                },
                "yearValue": {
                    "type": "string",
                    "pattern": "^\\d{4}$"
                },
                "monthValue": {
                    "type": "string",
                    "pattern": "^\\d{2}$"
                },
                "dayValue": {
                    "type": "string",
                    "pattern": "^\\d{2}$"
                },
                "zipcodeValue": {
                    "type": "string",
                    "pattern": "^\\d{5}$"
                }
            },
            "maxProperties": 2
        },
        "answer": {
            "type": "object",
            "properties": {
                "multipleIndex": {
                    "type": "integer",
                    "minimum": 0
                },
                "choice": {
                    "type": "integer",
                    "minimum": 1
                },
                "choices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/choicesItem"
                    },
                    "minItems": 1
                },
                "textValue": {
                    "type": "string"
                },
                "boolValue": {
                    "type": "boolean"
                },
                "dateValue": {
                    "type": "string",
                    "format": "date"
                },
                "numberValue": {
                    "type": "number"
                },
                "integerValue": {
                    "type": "integer"
                },
                "floatValue": {
                    "type": "number"
                },
                "feetInchesValue": {
                    "type": "object",
                    "properties": {
                        "feet": {
                            "type": "integer"
                        },
                        "inches": {
                            "type": "integer"
                        }
                    },
                    "minProperties": 1,
                    "additionalProperties": false
                },
                "bloodPressureValue": {
                    "type": "object",
                    "required": ["systolic", "diastolic"],
                    "properties": {
                        "systolic": {
                            "type": "integer"
                        },
                        "diastolic": {
                            "type": "integer"
                        }
                    },
                    "additionalProperties": false
                },
                "fileValue": {
                    "type": "object",
                    "properties": {
                        "id": {
                            "type": "integer",
                            "minimum": 1
                        },
                        "name": {
                            "type": "string",
                            "minLength": 1
                        },
                        "content": {
                            "type": "string",
                            "minLength": 1,
                            "description": "Base64 encoded binary or text content"
                        }
                    },
                    "minProperties": 1,
                    "maxProperties": 2
                },
                "yearValue": {
                    "type": "string",
                    "pattern": "^\\d{4}$"
                },
                "monthValue": {
                    "type": "string",
                    "pattern": "^\\d{2}$"
                },
                "dayValue": {
                    "type": "string",
                    "pattern": "^\\d{2}$"
                },
                "zipcodeValue": {
                    "type": "string",
                    "pattern": "^\\d{5}$"
                }
            },
            "minProperties": 1,
            "maxProperties": 2,
            "additionalProperties": false
        },
...
        "answerStatus": {
            "type": "string",
            "enum": ["new", "in-progress", "completed"]
        },
...
        "prevAssessment": {
            "type": "object",
            "required": ["prevAssessmentId"],
            "properties": {
                "prevAssessmentId": {
                    "type": "integer",
                    "minimum": 1
                }
            },
            "additionalProperties": false
        },
        "newComment": {
            "type": "object",
            "required": [
                "reason"
            ],
            "properties": {
                "reason": {
                    "type": "string",
                    "enum": ["agree", "disagree"]
                },
                "text": {
                    "type": "string"
                }
            },
            "additionalProperties": false
        },
        "newAssessmentLinkedAnswer": {
            "type": "object",
            "required": [
                "questionId"
            ],
            "properties": {
                "questionId": {
                    "type": "integer",
                    "minimum": 1
                },
                "language": {
                    "type": "string",
                    "minLength": 2
                },
                "meta": {
                    "type": "object"
                },
                "answer": {
                    "$ref": "#/definitions/answer"
                },
                "answers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/answer"
                    }
                },
                "comment": {
                    "$ref": "#/definitions/newComment"
                }
            },
            "additionalProperties": false
        },
        "newAssessmentLinkedAnswers": {
            "type": "array",
            "items": {
                "$ref": "#/definitions/newAssessmentLinkedAnswer"
            }
        },
        "newAssessmentAnswers": {
            "type": "object",
            "required": ["answers"],
            "properties": {
                "status": {
                    "$ref": "#/definitions/answerStatus"
                },
                "answers": {
                    "$ref": "#/definitions/newAssessmentLinkedAnswers"
                },
                "language": {
                    "type": "string",
                    "minLength": 2
                }
            },
            "additionalProperties": false
        },
        "comment": {
            "type": "object",
            "required": [
                "reason"
            ],
            "properties": {
                "id": {
                    "type": "integer",
                    "minimum": 1
                },
                "reason": {
                    "type": "string",
                    "enum": ["agree", "disagree"]
                },
                "text": {
                    "type": "string"
                },
                "userId": {
                    "type": "integer",
                    "minimum": 1
                },
                "language": {
                    "type": "string",
                    "minLength": 2
                }
            },
            "additionalProperties": false
        },
        "assessmentLinkedAnswer": {
            "type": "object",
            "required": [
                "questionId"
            ],
            "properties": {
                "questionId": {
                    "type": "integer",
                    "minimum": 1
                },
                "language": {
                    "type": "string",
                    "minLength": 2
                },
                "meta": {
                    "type": "object"
                },
                "answer": {
                    "$ref": "#/definitions/answer"
                },
                "answers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/answer"
                    }
                },
                "comment": {
                    "$ref": "#/definitions/comment"
                },
                "commentHistory": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/comment"
                    }
                }
            },
            "additionalProperties": false
        },
        "assessmentLinkedAnswers": {
            "type": "array",
            "items": {
                "$ref": "#/definitions/assessmentLinkedAnswer"
            }
        },
        "assessmentAnswers": {
            "type": "object",
            "required": ["status", "answers"],
            "properties": {
                "status": {
                    "$ref": "#/definitions/answerStatus"
                },
                "answers": {
                    "$ref": "#/definitions/assessmentLinkedAnswers"
                }
            },
            "additionalProperties": false
        },
        "assessmentAnswersInfo": {
            "type": "object",
            "required": ["id", "name", "status"],
            "properties": {
                "id": {
                    "type": "integer",
                    "minimum": 1
                },
                "name": {
                    "type": "string",
                    "minLength": 1
                },
                "stage": {
                    "type": "integer"
                },
                "group": {
                    "type": "string",
                    "minLength": 1
                },
                "status": {
                    "$ref": "#/definitions/answerStatus"
                }
            },
            "additionalProperties": false
        },
        "assessmentAnswersExportList": {
            "type": "array",
            "items": {
                "$ref": "#/definitions/assessmentAnswersExport"
            }
        },
        "assessmentAnswersExport": {
            "type": "object",
            "required": [
                "surveyId",
                "assessmentId",
                "userId",
                "questionId",
                "questionType",
                "value"

            ],
            "properties": {
                "surveyId": {
                    "type": "integer",
                    "minimum": 1
                },
                "surveyName": {
                    "type": "string",
                    "minimum": 1
                },
                "meta": {
                    "type": "object"
                },
                "assessmentId": {
                    "type": "integer",
                    "minimum": 1
                },
                "userId": {
                    "type": "integer",
                    "minimum": 1
                },
                "questionId": {
                    "type": "integer",
                    "minimum": 1
                },
                "questionType": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                },
                "questionText": {
                    "type":"string",
                    "minimum": 1
                },
                "questionInstruction": {
                    "type":"string",
                    "minimum": 1
                },
                "date": {
                    "type": "string",
                    "format": "dateTime",
                    "minimum": 1
                },
                "group": {
                    "type": "string",
                    "minimum": 1
                },
                "stage": {
                    "type": "string",
                    "minimum": 1
                },
                "code": {
                    "type": "string"
                }

            }
        },
        "assessmentAnswersDashboard": {
            "type": "array",
            "items": {
                "$ref": "#/definitions/assessmentAnswersInfo"
            },
            "additionalProperties": false
        },
...
    }
}